LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/instantiation-error-3-expected.txt

CONSOLE MESSAGE: SyntaxError: Indirectly exported binding name 'x' is not found.
CONSOLE MESSAGE: SyntaxError: Indirectly exported binding name 'x' is not found.
CONSOLE MESSAGE: SyntaxError: Indirectly exported binding name 'x' is not found.

FAIL Test that unresolvable cycles lead to SyntaxError events on window and load events on script null is not an object (evaluating 'ev.error.constructor')
